# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Franchise Generation API
  description: |
    REST API for the Franchise Generation platform — a multi-brand franchise ecosystem
    connecting brands, candidates, investors, consultants, and real estate developers.

    ## Authentication
    All protected endpoints require a JWT Bearer token obtained from `POST /auth/login`.
    Include the token in the `Authorization` header:
    ```
    Authorization: Bearer <token>
    ```

    ## Base URL
    Production: `https://api.franchisegeneration.be/api/v1`
    Local dev: `http://localhost:3000/api/v1`

    ## Roles
    | Role | Description |
    |------|-------------|
    | `admin` | Full backoffice access |
    | `investor` | Investment portal access |
    | `candidate` | Candidate onboarding portal |
    | `consultant` | Lead management access |
    | `developer` | Real estate developer access |
    | `brand` | Brand management portal |

  version: 1.0.0
  contact:
    name: Franchise Generation
    url: https://franchisegeneration.be
servers:
  - url: /api/v1
    description: Current version
security:
  - BearerAuth: []
...
paths:
...
  /support/tickets:
    get:
      tags: [Support]
      summary: List support tickets
      description: Returns all support tickets for the authenticated user. Admins see all tickets.
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [open, resolved, awaiting-investor, awaiting-admin]
        - name: category
          in: query
          schema:
            type: string
        - name: priority
          in: query
          schema:
            type: string
            enum: [low, normal, high, urgent]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Paginated list of tickets
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedList'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/SupportTicket'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Support]
      summary: Create support ticket
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [categoryId, subject]
              properties:
                categoryId:
                  type: string
                  enum: [investment, repayment, technical, document, legal, opportunity, general]
                  example: repayment
                priorityId:
                  type: string
                  enum: [low, normal, high, urgent]
                  default: normal
                subject:
                  type: string
                  example: "Question sur le remboursement de mai"
                brandId:
                  type: string
                  nullable: true
                  example: atelier
                message:
                  type: string
                  description: Body of the first message in the thread
                  example: "Bonjour, je n'ai pas reçu mon versement du 15 mai..."
            example:
              categoryId: repayment
              priorityId: normal
              subject: "Versement de mai non reçu"
              brandId: atelier
              message: "Bonjour, le versement prévu le 15 mai n'est pas arrivé sur mon compte."
      responses:
        '201':
          description: Ticket created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SupportTicket'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /support/tickets/{id}:
    get:
      tags: [Support]
      summary: Get ticket by ID
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: Ticket detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SupportTicket'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Support]
      summary: Update ticket
      description: Change ticket status or priority. Admins can reassign tickets.
      parameters:
        - $ref: '#/components/parameters/id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [open, resolved, awaiting-investor, awaiting-admin]
                priorityId:
                  type: string
                  enum: [low, normal, high, urgent]
                assignedTo:
                  type: string
                  description: Name of admin/agent assigned (admin only)
      responses:
        '200':
          description: Ticket updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SupportTicket'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /support/tickets/{id}/messages:
    get:
      tags: [Support]
      summary: Get ticket messages
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: Array of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SupportMessage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    post:
      tags: [Support]
      summary: Post message to ticket
      parameters:
        - $ref: '#/components/parameters/id'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [content]
              properties:
                content:
                  type: string
                  example: "Bonjour, j'ai déposé le document dans mon profil."
                attachments:
                  type: array
                  items:
                    type: string
                    format: binary
                  description: Optional file attachments (max 10 MB each)
      responses:
        '201':
          description: Message posted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SupportMessage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    id:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Resource identifier
...
  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: VALIDATION_ERROR
            message: Request validation failed.
            details:
              - field: email
                message: Must be a valid email address.
    Unauthorized:
      description: Missing or invalid authentication token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: UNAUTHORIZED
            message: A valid Bearer token is required.
    Forbidden:
      description: Authenticated but insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: FORBIDDEN
            message: You do not have permission to access this resource.
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: NOT_FOUND
            message: The requested resource was not found.
...
  schemas:
...
    SupportTicket:
      type: object
      description: A helpdesk support ticket
      properties:
        id:
          type: string
          example: t-001
        userId:
          type: string
        categoryId:
          type: string
          enum: [investment, repayment, technical, document, legal, opportunity, general]
          example: document
        priorityId:
          type: string
          enum: [low, normal, high, urgent]
          example: high
        subject:
          type: string
          example: "Justificatif de domicile à renouveler"
        status:
          type: string
          enum: [open, resolved, awaiting-investor, awaiting-admin]
          example: awaiting-investor
        brandId:
          type: string
          nullable: true
        assignedTo:
          type: string
          example: "Émilie Devos · Conformité"
        origin:
          type: string
          enum: [investor, admin, system]
          example: admin
        unreadMessages:
          type: integer
          example: 1
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        messages:
          type: array
          items:
            $ref: '#/components/schemas/SupportMessage'

    # -------------------------------------------------------------------------
    SupportMessage:
      type: object
      description: A message within a support ticket thread
      properties:
        id:
          type: string
          example: m-001
        ticketId:
          type: string
        fromType:
          type: string
          enum: [investor, admin, system]
          example: admin
        authorName:
          type: string
          example: "Émilie Devos"
        content:
          type: string
          example: "Bonjour Claire, votre justificatif de domicile date de plus de 2 ans..."
        unread:
          type: boolean
          example: true
        createdAt:
          type: string
          format: date-time
        attachments:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
              filename:
                type: string
              url:
                type: string
                format: uri
              size:
                type: string

    # -------------------------------------------------------------------------
...
    ErrorResponse:
      type: object
      description: Standard error response envelope
      properties:
        error:
          type: string
          description: Machine-readable error code
          example: VALIDATION_ERROR
        message:
          type: string
          description: Human-readable error description
          example: Request validation failed.
        details:
          type: array
          nullable: true
          description: Field-level validation errors
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string

    # -------------------------------------------------------------------------
    PaginatedList:
      type: object
      description: Generic paginated list wrapper
      properties:
        data:
          type: array
          items: {}
          description: Array of result items
        total:
          type: integer
          description: Total number of matching records
          example: 42
        page:
          type: integer
          description: Current page number (1-indexed)
          example: 1
        limit:
          type: integer
          description: Number of items per page
          example: 20
        totalPages:
          type: integer
          description: Total number of pages
          example: 3